Chatter/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
Flask==3.1.0
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.3
pandas==2.2.3
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
gunicorn==21.2.0
python-dotenv